+++ tests/test_utils.py
@@ -20,6 +20,7 @@ from datasets import load_dataset
 from transformers import AutoModelForCausalLM, AutoTokenizer, GenerationConfig
 from transformers.testing_utils import require_peft
 from transformers.utils import is_peft_available
+from parameterized import parameterized
 
 from trl import ModelConfig
 from trl.trainer import compute_accuracy
@@ -32,6 +33,7 @@ from trl.trainer.utils import (
     generate_model_card,
     get_peft_config,
     pad,
+    selective_log_softmax
 )
 
 
@@ -506,3 +508,32 @@ class TestComputeTokenAccuracy(unittest.TestCase):
         )
         accuracy = compute_token_accuracy(logits, labels)
         self.assertAlmostEqual(accuracy, 0.8)
+
+
+
+class TestSelectiveLogSoftmax(unittest.TestCase):
+    @parameterized.expand([
+        (torch.float64,),
+        (torch.float32,),
+        (torch.float16,),
+        (torch.bfloat16,)
+    ])
+    def test_selective_log_softmax(self, dtype):
+        """Test selective_log_softmax with logits of different dtypes"""
+        vocab_size = 32768
+        batch_size = 4
+        seq_len = 256
+
+        input_ids = torch.randint(low=0, high=vocab_size, size=(batch_size, seq_len), device="cuda")
+        logits = torch.randn(batch_size, seq_len, vocab_size, device="cuda", dtype=dtype)
+
+        expected_output = torch.gather(logits.log_softmax(-1), dim=-1, index=input_ids.unsqueeze(-1)).squeeze(
+            -1
+        )
+        actual_output = selective_log_softmax(logits=logits, input_ids=input_ids)
+
+        if dtype in [torch.float16, torch.bfloat16]:
+            # half-precision dtypes fall back to an exact method
+            self.assertTrue(torch.equal(actual_output, expected_output))
+        else:
+            torch.testing.assert_close(actual_output, expected_output, rtol=1e-5, atol=1e-5)
