+++ trl/trainer/nash_md_trainer.py
@@ -34,7 +34,6 @@ from transformers.trainer_utils import EvalPrediction
 from transformers.training_args import OptimizerNames
 from transformers.utils import is_apex_available
 
-from ..core import selective_log_softmax
 from ..data_utils import is_conversational, maybe_apply_chat_template
 from ..models.modeling_base import GeometricMixtureWrapper
 from ..models.utils import unwrap_model_for_generation
@@ -48,6 +47,7 @@ from .utils import (
     get_comet_experiment_url,
     get_reward,
     truncate_right,
+    selective_log_softmax
 )
