+++ trl/trainer/utils.py
@@ -1668,3 +1668,28 @@ def compute_token_accuracy(logits: torch.Tensor, labels: torch.Tensor, ignore_in
     accuracy = correct_tokens.item() / total_tokens.item() if total_tokens > 0 else 0.0
 
     return accuracy
+
+
+def selective_log_softmax(logits, input_ids):
+    """
+    A memory efficient implementation of a common log_softmax -> gather operation.
+
+    Equivalent to the following naive implementation:
+    ```python
+    per_token_logps = torch.gather(logits.log_softmax(-1), dim=-1, index=labels.unsqueeze(-1)).squeeze(-1)
+    ```
+    """
+    if logits.dtype in [torch.float32, torch.float64]:
+        selected_logits = torch.gather(logits, dim=-1, index=input_ids.unsqueeze(-1)).squeeze(-1)
+        # loop to reduce peak mem consumption
+        logsumexp_values = torch.stack([torch.logsumexp(lg, dim=-1) for lg in logits])
+        per_token_logps = selected_logits - logsumexp_values  # log_softmax(x_i) = x_i - logsumexp(x)
+    else:
+        # logsumexp approach is unstable with bfloat16, fall back to slightly less efficent approach
+        per_token_logps = []
+        for row_logits, row_labels in zip(logits, input_ids):  # loop to reduce peak mem consumption
+            row_logps = F.log_softmax(row_logits, dim=-1)
+            row_per_token_logps = row_logps.gather(dim=-1, index=row_labels.unsqueeze(-1)).squeeze(-1)
+            per_token_logps.append(row_per_token_logps)
+        per_token_logps = torch.stack(per_token_logps)
+    return per_token_logps
